+++ src/main/java/io/suricate/monitoring/services/nashorn/tasks/NashornRequestWidgetExecutionAsyncTask.java
@@ -159,7 +159,7 @@ public class NashornRequestWidgetExecutionAsyncTask implements Callable<NashornR
                     LOGGER.debug("The JSON response obtained after the execution of the Nashorn request of the widget instance {} is invalid", nashornRequest.getProjectWidgetId());
                     LOGGER.debug("The JSON response is: {}", json);
 
-                    nashornResponse.setLog(ToStringUtils.hideWidgetConfigurationInLogs(sw.toString() + "\nThe JSON response is not valid - " + json, widgetProperties.values()));
+                    nashornResponse.setLog(ToStringUtils.hideWidgetConfigurationInLogs(sw + "\nThe JSON response is not valid - " + json, widgetProperties.values()));
                     nashornResponse.setError(nashornRequest.isAlreadySuccess() ? NashornErrorTypeEnum.ERROR : NashornErrorTypeEnum.FATAL);
                 }
             }
@@ -179,7 +179,7 @@ public class NashornRequestWidgetExecutionAsyncTask implements Callable<NashornR
                     nashornResponse.setError(NashornErrorTypeEnum.ERROR);
                 }
 
-                nashornResponse.setLog(ExceptionUtils.getRootCauseMessage(exception));
+                nashornResponse.setLog(ExceptionUtils.getRootCause(exception).getMessage());
             }
         } finally {
             nashornResponse.setProjectId(nashornRequest.getProjectId());