+++ CHANGELOG.md
@@ -9,11 +9,10 @@
 - Refactor Admin menu
 - Immediately display the modifications when a widget has been edited
 - Improve search bar layout for admins
+- Improve widgets error logs
 
 🐛 Bugs fixing:
 
 - When backing to a previous step when adding a widget to a dashboard, do not display the input in error anymore
 
-📰 Other:
 
-- Docker image